split apart examples
